+++ CHANGELOG.md
@@ -1,2 +1,6 @@
 # OB-USP-AGENT Changelog since Release 8.0.0
 
+## 2024-04-15 v8.0.1
+### Fixed
+- USP Broker should not assume hierarchically ordered fields in GSDM and get instances responses
+- STOMP connects shouldn't block the data model thread